+++ apps/openmw/mwmechanics/pathfinding.hpp
@@ -51,7 +51,7 @@ namespace MWMechanics
                 return !mPath.empty();
             }
 
-            int getPathSize() const
+            size_t getPathSize() const
             {
                 return mPath.size();
             }
